git-svn-id: file:///tmp/svn_raytrace@148 10c6a05b-4505-4ba0-9d8c-39e3a15555b3

+++ Graphics/RayTracedImage.h
@@ -20,7 +20,7 @@ public:
 protected:
 	int height, width, size;
 	double pixelSize;
-	GLubyte *pixel;
+	GLdouble *pixel;
 };
 
 #endif /* RAYTRACEDIMAGE_H_ */
